fix: add 6492 verification  support to lab by using viem
