+++ apps/laboratory/src/utils/SignatureUtil.ts
@@ -1,0 +1,31 @@
+import { createPublicClient, http } from 'viem'
+
+function getTransport({ chainId }: { chainId: number }) {
+  const RPC_URL = 'https://rpc.walletconnect.com'
+
+  return http(
+    `${RPC_URL}/v1/?chainId=eip155:${chainId}&projectId=${process.env['NEXT_PUBLIC_PROJECT_ID']}`
+  )
+}
+
+export async function verifySignature({
+  address,
+  message,
+  signature,
+  chainId
+}: {
+  address: string
+  message: string
+  signature: string
+  chainId: number
+}) {
+  const publicClient = createPublicClient({
+    transport: getTransport({ chainId })
+  })
+
+  return publicClient.verifyMessage({
+    message,
+    address: address as `0x${string}`,
+    signature: signature as `0x${string}`
+  })
+}
